requirements.txt
transformers[torch]
torchvision
Pillow
torchmetrics
tqdm

